renamed index to sanitize
